requirements.txt
Flask
SQLAlchemy
psycopg2-binary
pandas
python-dotenv
ucimlrepo
scikit-learn
numpy
matplotlib
seaborn